+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/empties/IEmptiesInOutProducer.java
@@ -1,4 +1,4 @@
-package de.metas.handlingunits.inout;
+package de.metas.handlingunits.empties;
 
 /*
  * #%L